Add a TS test
