--- test/javascript/features/yoshi-server/local-middleware-route/route.test.ts
@@ -1,16 +1,0 @@
-import Scripts from '../../../../scripts';
-
-const scripts = Scripts.setupProjectFromTemplate({
-  templateDir: __dirname,
-  projectType: 'yoshi-server-javascript',
-});
-
-describe.each(['prod', 'dev'] as const)('yoshi-server route [%s]', mode => {
-  it('run tests', async () => {
-    await scripts[mode](async () => {
-      await page.goto(`${scripts.serverUrl}/app`);
-      const title = await page.$eval('h1', elm => elm.innerHTML);
-      expect(title).toBe('hello from my local middlware');
-    });
-  });
-});
