+++ test/typescript/features/yoshi-server/local-middleware-route/src/component.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function() {
+  return <div></div>;
+}
